+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KEZXQGSJKEWPMQEE8TS10177"
+        "01KF08B1ZJ9FHH4EY7F5940DT5"
     )
     r"""The unique identifier of the data source"""