+++ packages/orq-rc/RELEASES.md
@@ -4679,3 +4679,13 @@ Based on:
 - [python v4.2.0-rc.30] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.30] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.30 - packages/orq-rc
+
+## 2026-01-15 07:21:15
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.685.3 (2.795.8) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.31] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.31] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.31 - packages/orq-rc